+++ test/test_sidecar_utils.cpp
@@ -16,7 +16,7 @@
  *
 */
 
-#include "custom_collision_algo.hpp"
+#include "test_sidecar_utils.hpp"
 
 #include <SFML/Graphics.hpp>
 #include <imgui.h>
